refactor(ghost): streamline prompt strategy logic
Simplified `ContextAnalyzer` by removing complex `UseCaseType` determination logic, now directly checking for user input.
Refactored `PromptStrategyManager` to always select `UserRequestStrategy`, removing the need for multiple distinct strategy classes and their selection logic.
Introduced `GhostStrategy` interface and `GhostXmlStrategy` as the primary concrete implementation.
Removed obsolete strategy files (`AutoTriggerStrategy`, `BasePromptStrategy`, `PromptStrategy.ts`) and their associated types.

BREAKING CHANGE: The internal prompt strategy management system has been significantly refactored. The `GhostStrategy` class is now an interface, with `GhostXmlStrategy` serving as the primary concrete implementation. The `PromptStrategyManager` logic has been simplified to always select `UserRequestStrategy`, and several strategy-specific classes and types have been removed. Code that directly interacted with the old strategy classes, manager internals, or relied on specific strategy behaviors may require updates.
