+++ src/services/ghost/PromptStrategyManager.ts
@@ -1,25 +1,18 @@
 import { GhostSuggestionContext } from "./types"
-import { PromptStrategy } from "./types/PromptStrategy"
+import { PromptGenerator } from "./types/PromptGenerator"
 import { ContextAnalyzer } from "./ContextAnalyzer"
-
-// Import all strategies
-import { UserRequestStrategy } from "./strategies/UserRequestStrategy"
-import { AutoTriggerStrategy } from "./strategies/AutoTriggerStrategy"
+import { UserRequestPromptGenerator } from "./strategies/UserRequestPromptGenerator"
 
 /**
  * Manages prompt strategies and selects the appropriate one based on context
  */
 export class PromptStrategyManager {
-	private strategies: PromptStrategy[]
 	private contextAnalyzer: ContextAnalyzer
 	private debug: boolean
 
 	constructor(options?: { debug: boolean }) {
 		this.debug = options?.debug ?? false
 		this.contextAnalyzer = new ContextAnalyzer()
-
-		// Register all strategies in priority order
-		this.strategies = [new UserRequestStrategy(), new AutoTriggerStrategy()]
 	}
 
 	/**
@@ -27,7 +20,7 @@ export class PromptStrategyManager {
 	 * @param context The suggestion context
 	 * @returns The selected strategy
 	 */
-	selectStrategy(context: GhostSuggestionContext): PromptStrategy {
+	selectStrategy(context: GhostSuggestionContext): PromptGenerator {
 		// Analyze context to understand the situation
 		const analysis = this.contextAnalyzer.analyze(context)
 
@@ -42,25 +35,7 @@ export class PromptStrategyManager {
 				isInlineEdit: analysis.isInlineEdit,
 			})
 		}
-
-		// Find the first strategy that can handle this context
-		for (const strategy of this.strategies) {
-			if (strategy.canHandle(context)) {
-				if (this.debug) {
-					console.log(
-						`[PromptStrategyManager] Selected strategy: ${strategy.name} for use case: ${analysis.useCase}`,
-					)
-				}
-				return strategy
-			}
-		}
-
-		// Fallback: return the last strategy (AutoTriggerStrategy)
-		const fallback = this.strategies[this.strategies.length - 1]
-		if (this.debug) {
-			console.log(`[PromptStrategyManager] Falling back to: ${fallback.name}`)
-		}
-		return fallback
+		return new UserRequestPromptGenerator()
 	}
 
 	/**
@@ -71,7 +46,7 @@ export class PromptStrategyManager {
 	buildPrompt(context: GhostSuggestionContext): {
 		systemPrompt: string
 		userPrompt: string
-		strategy: PromptStrategy
+		strategy: PromptGenerator
 	} {
 		const strategy = this.selectStrategy(context)
 
@@ -92,13 +67,5 @@ export class PromptStrategyManager {
 			userPrompt,
 			strategy,
 		}
-	}
-
-	/**
-	 * Gets all registered strategies (for testing/debugging)
-	 * @returns Array of all strategies
-	 */
-	getStrategies(): PromptStrategy[] {
-		return [...this.strategies]
 	}
 }
