+++ src/services/ghost/__tests__/GhostXMLStrategy.spec.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from "vitest"
 import * as vscode from "vscode"
-import { GhostStrategy } from "../GhostStrategy"
+import { GhostXmlStrategy } from "../GhostXmlStrategy"
 import { GhostSuggestionContext, ASTContext } from "../types"
 import { MockTextDocument } from "../../mocking/MockTextDocument"
-import { StreamingParseResult } from "../GhostStreamingParser"
+import { PromptStrategyManager } from "../PromptStrategyManager"
 
 // Create a mock Node class for testing
 class MockNode {
@@ -105,6 +105,12 @@ vi.mock("vscode", () => ({
 			public end: any,
 		) {}
 	},
+	DiagnosticSeverity: {
+		Error: 0,
+		Warning: 1,
+		Information: 2,
+		Hint: 3,
+	},
 	workspace: {
 		asRelativePath: vi.fn().mockImplementation((uri) => {
 			if (typeof uri === "string") {
@@ -168,14 +174,14 @@ vi.mock("diff", () => ({
 	}),
 }))
 
-describe("GhostStrategy", () => {
-	let strategy: GhostStrategy
+describe("GhostXMLStrategy", () => {
+	let strategy: GhostXmlStrategy
 	let mockDocument: MockTextDocument
 	let mockASTContext: ASTContext
 	let mockRangeASTNode: MockNode
 
 	beforeEach(() => {
-		strategy = new GhostStrategy()
+		strategy = new GhostXmlStrategy()
 		mockDocument = new MockTextDocument(vscode.Uri.parse("file:///test.js"), "function test() {\n  return true;\n}")
 
 		// Create child node
@@ -357,6 +363,22 @@ describe("GhostStrategy", () => {
 			const completedChanges = strategy.getStreamingCompletedChanges()
 			expect(completedChanges.length).toBeGreaterThan(0)
 		})
+
+		it("should finish streaming parser and return final results", () => {
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+			}
+
+			strategy.initializeStreamingParser(context)
+
+			// Add some partial content
+			strategy.processStreamingChunk("<change><search><![CDATA[test")
+
+			// Finish should handle any remaining content
+			const result = strategy.finishStreamingParser()
+			expect(result).toBeDefined()
+			expect(result.suggestions).toBeDefined()
+		})
 	})
 
 	describe("prompt generation", () => {
@@ -379,6 +401,230 @@ describe("GhostStrategy", () => {
 			const suggestionPrompt = strategy.getSuggestionPrompt(context)
 			expect(suggestionPrompt).toContain("Add a comment")
 			expect(suggestionPrompt).toContain("<<<AUTOCOMPLETE_HERE>>>")
+		})
+
+		it("should use PromptStrategyManager to generate system prompt", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;",
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+				userInput: "Complete this function",
+			}
+
+			const prompt = strategy.getSystemPrompt(context)
+
+			// Should contain base instructions from strategy system
+			expect(prompt).toContain("CRITICAL OUTPUT FORMAT")
+		})
+
+		it("should select UserRequestStrategy when user input is provided", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;",
+				lineAt: (line: number) => ({ text: "const x = 1;" }),
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+				userInput: "Add a function to calculate sum",
+			}
+
+			const systemPrompt = strategy.getSystemPrompt(context)
+			const userPrompt = strategy.getSuggestionPrompt(context)
+
+			// UserRequestStrategy should be selected
+			expect(systemPrompt).toContain("Execute User's Explicit Request")
+			expect(userPrompt).toContain("Add a function to calculate sum")
+		})
+
+		it("should delegate to PromptStrategyManager", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;",
+				lineAt: (line: number) => ({ text: "const x = 1;" }),
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+			}
+
+			const prompt = strategy.getSuggestionPrompt(context)
+
+			// Should return a structured prompt
+			expect(prompt).toBeDefined()
+			expect(prompt.length).toBeGreaterThan(0)
+		})
+	})
+
+	describe("Integration", () => {
+		it("should work with both system and user prompts", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;",
+				lineAt: (line: number) => ({ text: "const x = 1;" }),
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+				userInput: "Complete this function",
+			}
+
+			const systemPrompt = strategy.getSystemPrompt(context)
+			const userPrompt = strategy.getSuggestionPrompt(context)
+
+			// System prompt should contain format instructions
+			expect(systemPrompt).toContain("CRITICAL OUTPUT FORMAT")
+			// User prompt should contain the user input
+			expect(userPrompt).toContain("Complete this function")
+		})
+
+		it("should properly delegate to PromptStrategyManager", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;",
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+			}
+
+			const systemPrompt = strategy.getSystemPrompt(context)
+			const userPrompt = strategy.getSuggestionPrompt(context)
+
+			expect(systemPrompt).toBeDefined()
+			expect(systemPrompt.length).toBeGreaterThan(0)
+			expect(userPrompt).toBeDefined()
+			expect(userPrompt.length).toBeGreaterThan(0)
+		})
+
+		it("should properly integrate with strategy manager for all use cases", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;\n",
+				lineAt: (line: number) => ({ text: line === 0 ? "const x = 1;" : "" }),
+				lineCount: 2,
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => (position.line === 0 ? 13 : 14),
+			} as vscode.TextDocument
+
+			// Test different contexts
+			const contexts = [
+				// User request
+				{
+					document: mockDocument,
+					userInput: "Add type annotation",
+				},
+				// Error fix
+				{
+					document: mockDocument,
+					diagnostics: [
+						{
+							severity: vscode.DiagnosticSeverity.Error,
+							message: "Type error",
+							range: { start: { line: 0 } } as vscode.Range,
+						} as vscode.Diagnostic,
+					],
+				},
+				// Selection refactor
+				{
+					document: mockDocument,
+					range: {
+						isEmpty: false,
+						start: { line: 0, character: 0 },
+						end: { line: 0, character: 12 },
+					} as vscode.Range,
+				},
+				// New line completion
+				{
+					document: mockDocument,
+					range: {
+						isEmpty: true,
+						start: { line: 1, character: 0 },
+					} as vscode.Range,
+				},
+			]
+
+			contexts.forEach((context) => {
+				const systemPrompt = strategy.getSystemPrompt(context as GhostSuggestionContext)
+				const userPrompt = strategy.getSuggestionPrompt(context as GhostSuggestionContext)
+
+				expect(systemPrompt).toBeDefined()
+				expect(systemPrompt.length).toBeGreaterThan(0)
+				expect(userPrompt).toBeDefined()
+				expect(userPrompt.length).toBeGreaterThan(0)
+			})
+		})
+	})
+
+	describe("Strategy Selection", () => {
+		let manager: PromptStrategyManager
+
+		beforeEach(() => {
+			manager = new PromptStrategyManager()
+		})
+
+		it("should select appropriate strategy based on context", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "// TODO: implement sum function\n",
+				lineAt: (line: number) => ({
+					text: line === 0 ? "// TODO: implement sum function" : "",
+					lineNumber: line,
+				}),
+				lineCount: 2,
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => (position.line === 1 ? 32 : 0),
+			} as vscode.TextDocument
+
+			const mockRange = {
+				start: { line: 1, character: 0 } as vscode.Position,
+				end: { line: 1, character: 0 } as vscode.Position,
+				isEmpty: true,
+			} as vscode.Range
+
+			// Test comment-driven context
+			const commentContext: GhostSuggestionContext = {
+				document: mockDocument,
+				range: mockRange,
+			}
+
+			const result = manager.buildPrompt(commentContext)
+
+			// Should select UserRequestStrategy when there's a comment context
+			// The strategy selection logic prioritizes user requests over new line completion
+			expect(result.strategy.name).toBe("User Request")
+		})
+
+		it("should fallback to UserRequestStrategy when no specific strategy matches", () => {
+			const mockDocument = {
+				languageId: "typescript",
+				getText: () => "const x = 1;\n",
+				lineAt: (line: number) => ({ text: "const x = 1;" }),
+				uri: { toString: () => "file:///test.ts" },
+				offsetAt: (position: vscode.Position) => 13,
+			} as vscode.TextDocument
+
+			const context: GhostSuggestionContext = {
+				document: mockDocument,
+			}
+
+			const result = manager.buildPrompt(context)
+
+			// Should fallback to UserRequestStrategy (the default strategy)
+			expect(result.strategy.name).toBe("User Request")
 		})
 	})
 })
